Release 0.1.57 (#106)

+++ src/common/load/loadApplication.ts
@@ -41,8 +41,10 @@ interface IParams {
 }
 
 async function preInit({ temporaryPath, applicationPath }) {
+  const hookPath = path.join(temporaryPath, 'hook');
+  if (!fs.existsSync(hookPath)) return;
   try {
-    const baseChildComponent = await require(path.join(temporaryPath, 'hook'));
+    const baseChildComponent = await require(hookPath);
     const tempObj = {
       tempPath: temporaryPath,
       targetPath: applicationPath,
@@ -162,28 +164,32 @@ class LoadApplication {
     return applicationPath;
   }
   async postInit({ temporaryPath, applicationPath, parameters }) {
+    const hookPath = path.join(temporaryPath, 'hook');
     let response: any = {};
-    try {
-      const baseChildComponent = await require(path.join(temporaryPath, 'hook'));
-      const tempObj = {
-        tempPath: temporaryPath,
-        targetPath: applicationPath,
-        downloadRequest: downloadRequest,
-        fse: fs,
-        lodash: _,
-        artTemplate: (filePath: string) => {
-          const newPath = path.join(applicationPath, filePath);
-          const newData = this.handleArtTemplate(newPath, parameters);
-          fs.writeFileSync(newPath, newData, 'utf-8');
-        },
-        Logger,
-        execCommand,
-        parameters,
-      };
-      response = await baseChildComponent.postInit(tempObj);
-    } catch (e) {
-      logger.debug(`postInit error: ${e}`);
+    if (fs.existsSync(hookPath)) {
+      try {
+        const baseChildComponent = await require(hookPath);
+        const tempObj = {
+          tempPath: temporaryPath,
+          targetPath: applicationPath,
+          downloadRequest: downloadRequest,
+          fse: fs,
+          lodash: _,
+          artTemplate: (filePath: string) => {
+            const newPath = path.join(applicationPath, filePath);
+            const newData = this.handleArtTemplate(newPath, parameters);
+            fs.writeFileSync(newPath, newData, 'utf-8');
+          },
+          Logger,
+          execCommand,
+          parameters,
+        };
+        response = await baseChildComponent.postInit(tempObj);
+      } catch (e) {
+        logger.debug(`postInit error: ${e}`);
+      }
     }
+    if (isEmpty(this.spath)) return;
     // _custom_secret_list：postInit 里面的 secret 字段
     const { _custom_secret_list, ...rest } = response || {};
     const result = {
